+++ lib/apickli-gherkin.js
@@ -1,6 +1,6 @@
 const { Given, When, Then } = require('cucumber')
 const { Async: Task, curry, propOr } = require('crocks')
-const apickli = require('./index.js')
+const apickli = require('./apickli.js')
 const util = require('./util.js')
 const fs = require('fs')
